// oasis/src/texas/types.gen.ts
export type ErrorResponse = {
  error: OAuthErrorCode;
  error_description: string;
};

/**
 * Identity providers for use with token fetch, exchange and introspection.
 */
export type IdentityProvider =
  | "azuread"
  | "tokenx"
  | "maskinporten"
  | "idporten";
// ...
export type OAuthErrorCode =
  | "invalid_request"
  | "invalid_client"
  | "invalid_grant"
  | "unauthorized_client"
  | "unsupported_grant_type"
  | "invalid_scope"
  | "server_error";
// ...
export type TokenRequest = {
  identity_provider: IdentityProvider;
  /**
   * Resource indicator for audience-restricted tokens (RFC 8707).
   */
  resource?: string | null;
  /**
   * Force renewal of token. Defaults to false if omitted.
   */
  skip_cache?: boolean | null;
  /**
   * Scope or identifier for the target application.
   */
  target: string;
};

/**
 * RFC 6749 token response from section 5.1.
 */
export type TokenResponse = {
  access_token: string;
  /**
   * Token expiry in seconds. Useful for caching purposes.
   */
  expires_in: number;
  token_type: TokenType;
};

/**
 * Token type is always Bearer, but this might change in the future.
 */
export type TokenType = "Bearer";